+++ src/main/java/com/example/gatherplan/appointment/service/TempAppointmentService.java
@@ -3,5 +3,5 @@ package com.example.gatherplan.appointment.service;
 import com.example.gatherplan.appointment.dto.CreateTempAppointmentReqDto;
 
 public interface TempAppointmentService {
-    String registerTempAppointment(CreateTempAppointmentReqDto createTempAppointmentReqDto);
+    String registerTempAppointment(CreateTempAppointmentReqDto reqDto);
 }
