+++ src/main/java/com/example/gatherplan/appointment/service/impl/UserServiceImpl.java
@@ -10,10 +10,10 @@ import com.example.gatherplan.appointment.repository.UserRepository;
 import com.example.gatherplan.appointment.repository.entity.EmailAuth;
 import com.example.gatherplan.appointment.repository.entity.User;
 import com.example.gatherplan.appointment.service.UserService;
+import com.example.gatherplan.common.config.jwt.RoleType;
+import com.example.gatherplan.common.config.jwt.UserInfo;
 import com.example.gatherplan.common.exception.AuthenticationFailException;
 import com.example.gatherplan.common.exception.ErrorCode;
-import com.example.gatherplan.common.jwt.RoleType;
-import com.example.gatherplan.common.jwt.UserInfo;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
