+++ src/main/java/com/example/gatherplan/common/config/SecurityConfig.java
@@ -1,9 +1,9 @@
 package com.example.gatherplan.common.config;
 
-import com.example.gatherplan.common.jwt.JWTFilter;
-import com.example.gatherplan.common.jwt.JWTUtil;
-import com.example.gatherplan.common.jwt.LoginFilter;
-import com.example.gatherplan.common.jwt.RoleType;
+import com.example.gatherplan.common.config.jwt.JWTFilter;
+import com.example.gatherplan.common.config.jwt.JWTUtil;
+import com.example.gatherplan.common.config.jwt.LoginFilter;
+import com.example.gatherplan.common.config.jwt.RoleType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.Filter;
 import lombok.RequiredArgsConstructor;
