+++ src/main/java/com/example/gatherplan/common/config/jwt/JWTFilter.java
@@ -1,4 +1,4 @@
-package com.example.gatherplan.common.jwt;
+package com.example.gatherplan.common.config.jwt;
 
 import com.example.gatherplan.appointment.repository.entity.User;
 import jakarta.servlet.FilterChain;
