+++ src/main/java/com/example/gatherplan/common/config/jwt/LoginReq.java
@@ -1,4 +1,4 @@
-package com.example.gatherplan.common.jwt;
+package com.example.gatherplan.common.config.jwt;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
