--- src/main/java/com/example/gatherplan/common/validation/SizeCheckAuthcode.java
@@ -1,4 +1,0 @@
-package com.example.gatherplan.common.validation;
-
-public interface SizeCheckAuthcode {
-}
