--- src/main/java/com/example/gatherplan/common/validation/SizeCheckNickName.java
@@ -1,4 +1,0 @@
-package com.example.gatherplan.common.validation;
-
-public interface SizeCheckNickName {
-}
