+++ src/main/java/com/example/gatherplan/controller/validation/CreateUserReqValidSeq.java
@@ -1,6 +1,5 @@
 package com.example.gatherplan.controller.validation;
 
-import com.example.gatherplan.common.validation.*;
 import jakarta.validation.GroupSequence;
 
 @GroupSequence({NotBlankEmail.class, PatternCheckEmail.class, NotBlankAuthcode.class, SizeCheckAuthcode.class
