+++ src/main/java/com/example/gatherplan/controller/validation/NotBlankAppointmentName.java
@@ -1,0 +1,4 @@
+package com.example.gatherplan.controller.validation;
+
+public interface NotBlankAppointmentName {
+}
