+++ src/main/java/com/example/gatherplan/controller/validation/PatternCheckEmail.java
@@ -1,0 +1,4 @@
+package com.example.gatherplan.controller.validation;
+
+public interface PatternCheckEmail {
+}
